bentobox: fix GDT & aarch64 build

+++ base/usr/include/kernel/arch/x86_64/gdt.h
@@ -36,5 +36,5 @@ struct gdtr {
 extern struct gdt_table gdt_table;
 
 void gdt_install(void);
-void gdt_set_entry(uint8_t index, uint16_t limit, uint64_t base, uint8_t access, uint8_t gran);
+void gdt_set_entry(uint16_t index, uint16_t limit, uint64_t base, uint8_t access, uint8_t gran);
 void gdt_flush(void);
