+++ kernel/arch/x86_64/gdt.c
@@ -6,7 +6,7 @@
 struct gdt_table gdt_table;
 struct gdtr gdt_descriptor;
 
-void gdt_set_entry(uint8_t index, uint16_t limit, uint64_t base, uint8_t access, uint8_t gran) {
+void gdt_set_entry(uint16_t index, uint16_t limit, uint64_t base, uint8_t access, uint8_t gran) {
     if (index > 4) {
         int i = index - 5;
         gdt_table.tss_entries[i].limit = limit;
